feat: Move did:orb to did:web transformation code to separate function
This function will be used by the client to verify that did:web has been created from did:orb

Closes #1465
